* general-fixes:
  gcc-external{.inc,-cross}: use non-prefixed gcc_binaries
  gcc-external.inc: drop unnecessary external-toolchain inherit
  layer.conf: add fallback default for GCC_VERSION
  gcc: Use globally set GCC_VERSION in recipes
  tcmode-external: also blacklist libgfortran
  tcmode-external: add prefs for gcc-runtime, libgcc-initial, libgfortran
  meta-environment{,-extsdk}: don't include -B${gcc_bindir} in emitted TUNE_CC_ARCH
  external-toolchain.bbclass: expand permissions beyond user
  libgcc-external: drop the static files that won't be found anyway
  libgfortran-external: don't package the static lib and headers
  external-toolchain-cross.bbclass: drop sourcery-specific nopseudo handling
  oe.external: obey EXTERNAL_FILES over FILES
  oe.external: drop unneeded ,True from getVars
  external-toolchain.bbclass: stop disabling strip
  external-toolchain.bbclass: use PACKAGE_DEPENDS
  binutils-external-cross: only link ld.bfd if necessary

+++ recipes-external/binutils/binutils-external-cross.bb
@@ -15,5 +15,7 @@ EXTERNAL_CROSS_BINARIES = "ar as ld nm objcopy objdump ranlib strip \
                            strings"
 
 do_install_append () {
-    ln -s ${TARGET_PREFIX}ld ${D}${bindir}/${TARGET_PREFIX}ld.bfd
+    if [ ! -e ${D}${bindir}/${TARGET_PREFIX}ld.bfd ]; then
+        ln -s ${TARGET_PREFIX}ld ${D}${bindir}/${TARGET_PREFIX}ld.bfd
+    fi
 }
